Column headers now far more robust to real tables
Also changed around some of the semantics to ColumnHeader: it was before giving
empty strings in one case and failing in another case.  I decided that empty
cases should be None for both the class and the text, though I'm not positive
that's semantically the right thing to do.  I'm going with it.