Patched a bug: new drafts documents now work.
Fixed a bug where sports-reference.com now adds a table to the bottom that
lists how draft picks evolved over time to be relevant to the draft.  Now, for
understanding purposes, this is obviously great data.  For my parser, it
assumes it's a data table, and gets confused when it doesn't find data or
headers on the plain text.  Therefore, for now, I filter it out.

If you find more cases of "table-less headers" present, you can add them to the
blacklisted table header section.  If I find one more, I'll refactor the code
to make it a bit more straightforward.  If I find a *lot* more, I'll have to
find a more programmatic way to find these.